# Conflicts:
#	.idea/workspace.xml
#	Files/config.json
#	out/production/dpoo-2021-smartpiano-ice5/Business/BusinessFacade.class
#	out/production/dpoo-2021-smartpiano-ice5/Business/BusinessFacadeImp.class
#	out/production/dpoo-2021-smartpiano-ice5/Business/Entities/MidiHelper.class
#	out/production/dpoo-2021-smartpiano-ice5/Business/Entities/Song.class
#	out/production/dpoo-2021-smartpiano-ice5/Business/PlaylistManager.class
#	out/production/dpoo-2021-smartpiano-ice5/Business/SongManager.class
#	out/production/dpoo-2021-smartpiano-ice5/Business/TilesManager.class
#	out/production/dpoo-2021-smartpiano-ice5/Business/TimerManager.class
#	out/production/dpoo-2021-smartpiano-ice5/Persistence/SQL/Csv/LoginUserCsvDAO.class
#	out/production/dpoo-2021-smartpiano-ice5/Persistence/SQL/Csv/PlaylistCsvDAO.class
#	out/production/dpoo-2021-smartpiano-ice5/Persistence/SQL/Csv/SongCsvDAO.class
#	out/production/dpoo-2021-smartpiano-ice5/Presentation/DictionaryPiano.class
#	out/production/dpoo-2021-smartpiano-ice5/Presentation/Manager/FreePianoUIManager$1.class
#	out/production/dpoo-2021-smartpiano-ice5/Presentation/Manager/FreePianoUIManager.class
#	out/production/dpoo-2021-smartpiano-ice5/Presentation/Manager/PianoFrameManager.class
#	out/production/dpoo-2021-smartpiano-ice5/Presentation/Manager/PianoTilesUISelectorManager$1.class
#	out/production/dpoo-2021-smartpiano-ice5/Presentation/Manager/PianoTilesUISelectorManager.class
#	out/production/dpoo-2021-smartpiano-ice5/Presentation/Manager/PreMenuUIManager.class
#	out/production/dpoo-2021-smartpiano-ice5/Presentation/Manager/ProfileUIManager.class
#	out/production/dpoo-2021-smartpiano-ice5/Presentation/Manager/SpotiFrameManager.class
#	out/production/dpoo-2021-smartpiano-ice5/Presentation/Ui_Views/PianoFrame.class
#	out/production/dpoo-2021-smartpiano-ice5/Presentation/Ui_Views/PianoTilesUISelector.class
#	out/production/dpoo-2021-smartpiano-ice5/Presentation/Ui_Views/PlaylistUI.class
#	out/production/dpoo-2021-smartpiano-ice5/Presentation/Ui_Views/SongsUI$1.class
#	out/production/dpoo-2021-smartpiano-ice5/Presentation/Ui_Views/SongsUI.class
#	out/production/dpoo-2021-smartpiano-ice5/Presentation/Ui_Views/SpotiUI.class
#	out/production/dpoo-2021-smartpiano-ice5/Presentation/Ui_Views/StatisticsUI.class
#	out/production/dpoo-2021-smartpiano-ice5/Presentation/Ui_Views/Tile.class
#	src/Business/BusinessFacadeImp.java
#	src/Business/Entities/MidiHelper.java
#	src/Business/Entities/Song.java
#	src/Business/SongManager.java
#	src/Persistence/SQL/Csv/SongCsvDAO.java
#	src/Presentation/Manager/ProfileUIManager.java
#	src/Presentation/Ui_Views/PlaylistUI.java
#	src/Presentation/Ui_Views/SpotiUI.java

+++ src/Persistence/SQL/Csv/SongCsvDAO.java
@@ -25,23 +25,34 @@ public class SongCsvDAO implements SongDAO {
      * @param myUserString Defines the username of the user who we want to get their songs
      * @return List of the class Song that stores the songs created by that User
      */
-    private ArrayList<Song> songFromCsv(String myUserString) throws SQLException{
+
+    private ArrayList<Song> songFromCsv(String myUserString) throws SQLException {
+        if(ConnectSQL.getInstance()!=null) {
             ResultSet myRs = ConnectSQL.getInstance().createStatement().executeQuery("select * from SongT as s where (s.username " +
-                    "                    like '"+ myUserString+"' and publicBoolean = false) or publicBoolean = true;");
+                    "                    like '" + myUserString + "' and publicBoolean = false) or publicBoolean = true;");
             ArrayList<Song> songs = myRsToSongs(myRs);
             myRs.close();
             return songs;
-
+        }
+        return null;
     }
 
-
-    private ArrayList<Song> songFromCsv() throws SQLException{
+    /**
+     * Method that gets all the songs from the database that are public
+     * @return ArrayList of songs
+     */
+    private ArrayList<Song> songFromCsv() throws SQLException {
+        if (ConnectSQL.getInstance() != null) {
             ResultSet myRs = ConnectSQL.getInstance().createStatement().executeQuery("select * from SongT as s where" +
                     " publicBoolean = true;");
             ArrayList<Song> songs = myRsToSongs(myRs);
             myRs.close();
             return songs;
+
+        }
+        return null;
     }
+
 
     /**
      * Method that parses the result got from the query and stores it in the list
@@ -69,12 +80,11 @@ public class SongCsvDAO implements SongDAO {
 
     /**
      * Method that stores the song in the database
-     *
      * @param mySaveSong Defines the song to be stored
      */
     @Override
     public boolean saveSong(Song mySaveSong) {
-        try {
+        if(ConnectSQL.getInstance()!=null){
             PreparedStatement st = ConnectSQL.getInstance().prepareStatement(
                     "insert into SongT (songName, authorsName, duration, recordingDate, publicBoolean, songFile, username) values ('" +
                             mySaveSong.getSongName() + "', '" +
@@ -86,25 +96,30 @@ public class SongCsvDAO implements SongDAO {
                             mySaveSong.getCreator() + "')");
             st.execute();
             return true;
-
-        } catch (SQLException throwable) {
-            return false;
         }
+        return false;
+    } catch (SQLException throwable) {
+        return false;
     }
+}
+
     public boolean saveSongWithDate(Song mySaveSong) {
         try {
-            PreparedStatement st = ConnectSQL.getInstance().prepareStatement(
-                    "insert into SongT (songName, authorsName, duration, recordingDate, publicBoolean, songFile, username, numTimesPlayed) values ('" +
-                            mySaveSong.getSongName() + "', '" +
-                            mySaveSong.getAuthorName() + "', '" +
-                            mySaveSong.getDuration() + "', '" +
-                            new SimpleDateFormat("yyyy-MM-dd").format(mySaveSong.getRecordingDate()) + "', " +
-                            mySaveSong.isPublicBoolean() + ", '" +
-                            mySaveSong.getSongFile() + "', '" +
-                            mySaveSong.getCreator() + "'," +  mySaveSong.getTimesPlayed() +
-                            ")");
-            st.execute();
-            return true;
+            if(ConnectSQL.getInstance()!=null) {
+                PreparedStatement st = ConnectSQL.getInstance().prepareStatement(
+                        "insert into SongT (songName, authorsName, duration, recordingDate, publicBoolean, songFile, username, numTimesPlayed) values ('" +
+                                mySaveSong.getSongName() + "', '" +
+                                mySaveSong.getAuthorName() + "', '" +
+                                mySaveSong.getDuration() + "', '" +
+                                new SimpleDateFormat("yyyy-MM-dd").format(mySaveSong.getRecordingDate()) + "', " +
+                                mySaveSong.isPublicBoolean() + ", '" +
+                                mySaveSong.getSongFile() + "', '" +
+                                mySaveSong.getCreator() + "'," + mySaveSong.getTimesPlayed() +
+                                ")");
+                st.execute();
+                return true;
+            }
+            return false;
         } catch (SQLException throwable) {
             return false;
         }
@@ -112,16 +127,18 @@ public class SongCsvDAO implements SongDAO {
 
     /**
      * Method that deletes the song depending on the id from the database
-     *
      * @param mySong Defines the song to be deleted from the database
      */
     @Override
     public boolean deleteSong(Song mySong) {
         try {
-            PreparedStatement st = ConnectSQL.getInstance().prepareStatement("delete from SongT where songName = '" +
-                    mySong.getSongName() + "'");
-            st.execute();
-            return true;
+            if(ConnectSQL.getInstance()!=null){
+                PreparedStatement st = ConnectSQL.getInstance().prepareStatement("delete from SongT where songName = '" +
+                        mySong.getSongName() + "'");
+                st.execute();
+                return true;
+            }
+            return false;
         } catch (SQLException e) {
             return false;
         }
@@ -136,25 +153,27 @@ public class SongCsvDAO implements SongDAO {
     @Override
     public Song getSongByID(int id) {
         try {
-            ResultSet myRs = ConnectSQL.getInstance().createStatement().executeQuery("select * from SongT as s where s.songId LIKE " + id);
-            if (myRs.next()) {
-                Song song = new Song(
-                        myRs.getString("songName"),
-                        myRs.getString("authorsName"),
-                        myRs.getFloat("duration"),
-                        myRs.getDate("recordingDate"),
-                        myRs.getBoolean("publicBoolean"),
-                        myRs.getString("songFile"),
-                        myRs.getString("username"),
-                        myRs.getInt("numTimesPlayed"));
-                myRs.close();
-                return song;
-            } else {
-                return null;
+            if(ConnectSQL.getInstance()!=null){
+
+                ResultSet myRs = ConnectSQL.getInstance().createStatement().executeQuery("select * from SongT as s where s.songId LIKE " + id);
+                if (myRs.next()) {
+                    Song song = new Song(
+                            myRs.getString("songName"),
+                            myRs.getString("authorsName"),
+                            myRs.getFloat("duration"),
+                            myRs.getDate("recordingDate"),
+                            myRs.getBoolean("publicBoolean"),
+                            myRs.getString("songFile"),
+                            myRs.getString("username"),
+                            myRs.getInt("numTimesPlayed"));
+                    myRs.close();
+                    return song;
+                }
             }
         } catch (SQLException throwables) {
             return null;
         }
+        return null;
 
     }
